% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayeslist-package.R
\docType{package}
\name{bayeslist-package}
\alias{bayeslist-package}
\alias{bayeslist}
\title{The 'bayeslist' package.}
\description{
Estimate list experiment data applying full Bayesian approaches.
}
\references{
Stan Development Team (2018). RStan: the R interface to Stan. R package version 2.18.2. https://mc-stan.org
}
