% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printfun.R
\name{coef.bayeslist}
\alias{coef.bayeslist}
\title{Extract coefficients from a \code{bayeslist} object}
\usage{
\method{coef}{bayeslist}(object, ...)
}
\arguments{
\item{object}{A \code{bayeslist} object from running the main function \code{\link{bayeslist}}.}

\item{...}{Further arguments to be passed according to \code{\link[stats]{coef}}.}
}
\value{
A table of coefficients with their corresponding lower and upper bounds.
}
\description{
Create a table of coefficient results from a \code{bayeslist} object.
}
