% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotfun.R
\name{logistic}
\alias{logistic}
\title{Logistic function}
\usage{
logistic(x)
}
\arguments{
\item{x}{A scalar or vector to be logit-transformed.}
}
\value{
logit-transformed value
}
\description{
Standard logistic function.
}
