% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotfun.R
\name{plot.bayeslist}
\alias{plot.bayeslist}
\title{Plot bayeslist object}
\usage{
\method{plot}{bayeslist}(x, type = "prevalence", ...)
}
\arguments{
\item{x}{A \code{bayeslist} object to be plotted.}

\item{type}{Character string giving the type of plotting. The options are
\code{"trace"} for trace plots, \code{"prevalence"} for prevalence plots. The default is "prevalence".}

\item{...}{Additional arguments to be passed to subsequent plot functions (check the See Also section).}
}
\value{
None.
}
\description{
General plot function for \code{bayeslist} objects, which dispatches the chosen
type of plotting to the corresponding function.
}
\seealso{
\code{\link{plot_trace.bayeslist}} and \code{\link{plot_coef.bayeslist}}.
}
