% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotfun.R
\name{plot_coef.bayeslist}
\alias{plot_coef.bayeslist}
\title{Make coefficient plots for a \code{bayeslist} object}
\usage{
plot_coef.bayeslist(object, ...)
}
\arguments{
\item{object}{A \code{bayeslist} object from running the main function \code{\link{bayeslist}}.}

\item{...}{Additional parameters to be passed to \code{\link[rstan]{stan_plot}}.}
}
\value{
None.
}
\description{
\code{plot_coef.bayeslist} is used to produce coefficient plots from a \code{bayeslist} object.
}
