% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotfun.R
\name{plot_trace.bayeslist}
\alias{plot_trace.bayeslist}
\title{Trace plots for bayeslist}
\usage{
plot_trace.bayeslist(object, ...)
}
\arguments{
\item{object}{A \code{bayeslist} object from running the main function \code{\link{bayeslist}}.}

\item{...}{Additional parameters to be passed to \code{\link[rstan]{traceplot}}.}
}
\value{
None.
}
\description{
\code{plot_trace.bayeslist} is used to produce trace plots from a \code{bayeslist} object from the main function \code{\link{bayeslist}}.
}
