% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printfun.R
\name{print.bayeslist}
\alias{print.bayeslist}
\title{Print returns from a \code{bayeslist} object}
\usage{
\method{print}{bayeslist}(x, type = "text", ...)
}
\arguments{
\item{x}{A \code{bayeslist} object to be printed.}

\item{type}{Character string giving the type of printing, such as
\code{"text"}, \code{"mcmc"}, \code{"coef"}.}

\item{...}{Additional arguments to be passed to print functions (check the See Also section).}
}
\value{
None.
}
\description{
General print function for \code{bayeslist} objects, which dispatches the chosen type
of printing to the corresponding function.
}
\seealso{
\code{\link{print_text.bayeslist}}, \code{\link{print_mcmc.bayeslist}}, \code{\link{print_coef.bayeslist}}.
}
