% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printfun.R
\name{print_coef.bayeslist}
\alias{print_coef.bayeslist}
\title{Print coefficients of a \code{bayeslist} object}
\usage{
print_coef.bayeslist(object, digits = 3)
}
\arguments{
\item{object}{A \code{bayeslist} object.}

\item{digits}{Number of digits to display.}
}
\value{
None.
}
\description{
\code{print_coef.bayeslist} prints out coefficients from a \code{bayeslist} object from running the main function \code{\link{bayeslist}}.
}
