% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printfun.R
\name{print_mcmc.bayeslist}
\alias{print_mcmc.bayeslist}
\title{Print convergence diagnostics from a \code{bayeslist} object}
\usage{
print_mcmc.bayeslist(object, ...)
}
\arguments{
\item{object}{A \code{bayeslist} object.}

\item{...}{Additional arguments to be passed to the \code{print} function.}
}
\value{
None.
}
\description{
\code{print_mcmc.bayeslist} prints a number of diagnostics about the convergence of a \code{bayeslist} objects.
}
