% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printfun.R
\name{summary.bayeslist}
\alias{summary.bayeslist}
\title{Summary of a \code{bayeslist} object}
\usage{
\method{summary}{bayeslist}(object, ...)
}
\arguments{
\item{object}{A \code{bayeslist} object to be summarized.}

\item{...}{Additional arguments to be passed to summary function.}
}
\value{
None.
}
\description{
General summary function for \code{bayeslist} objects.
}
\seealso{
\code{\link{print_text.bayeslist}}, \code{\link{print_mcmc.bayeslist}}, \code{\link{print_coef.bayeslist}}.
}
