% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayeslist-package.R
\docType{package}
\name{bayeslist-package}
\alias{bayeslist-package}
\title{The 'bayeslist' package.}
\description{
Estimate list experiment data applying full Bayesian approaches.
}
\author{
\strong{Maintainer}: Xiao Lu \email{xiao.lu.research@gmail.com}

Authors:
\itemize{
  \item Richard Traunmüller
}

}
